Made improvements to collision detection using newly implemented Event and EventManager classes.
Implemented a number of changes to the project overall as instructed in the "Bits" assignment.

These changes include:
Two new methods to "get" Actors in a scene, using the Name and Tag respectively. Getting from tag returns a list of Actors.
Improvements to JSON data parsing to reduce errors.
Three new functions so that position, rotation, and scale can be decomposed from the 3x3 Matrix.
Clamp, lerp, normalise, remap, and wrap functions implemented in MathUtils.
Made the engine automatically detect the collision box for an Actor based on Sprite size.